requirements.txt
Django
django-ckeditor
python-dotenv
django-jazzmin